Stories migrated from firebase to php backend

+++ app/src/main/java/com/example/a22i1066_b_socially/PostAdapter.kt
@@ -78,7 +78,7 @@ class PostAdapter(
         notifyDataSetChanged()
     }
 
-    fun updatePost(postId: String, newLikesCount: Long, isLiked: Boolean) {
+    fun updatePost(postId: String, newLikesCount: Int, isLiked: Boolean) {
         val index = posts.indexOfFirst { it.id == postId }
         if (index != -1) {
             posts[index].likesCount = newLikesCount
